refactor(ui): simplify list components and remove item_label parameter
- Remove item_label parameter from SearchList and Typeahead as it's handled by ItemList
- Rename BaseList to ItemList for better clarity
- Clean up component inheritance to reduce duplication
